// src/generated/types.gen.ts
export type ResourceIdentifier = {
    /**
     * The unique id of the referenced resource
     */
    rid: string;
    /**
     * The type of the referenced resource
     */
    rtype: 'device' | 'bridge_home' | 'room' | 'zone' | 'light' | 'button' | 'relative_rotary' | 'temperature' | 'light_level' | 'motion' | 'camera_motion' | 'entertainment' | 'contact' | 'tamper' | 'grouped_light' | 'device_power' | 'zigbee_bridge_connectivity' | 'zigbee_connectivity' | 'zgp_connectivity' | 'bridge' | 'zigbee_device_discovery' | 'homekit' | 'matter' | 'matter_fabric' | 'scene' | 'entertainment_configuration' | 'public_image' | 'auth_v1' | 'behavior_script' | 'behavior_instance' | 'geofence' | 'geofence_client' | 'geolocation' | 'smart_scene' | 'grouped_motion' | 'grouped_light_level';
};
// ...
export type On = {
    /**
     * On/Off state of the light on=true, off=false
     */
    on: boolean;
};
// ...
export type Color = {
    xy: GamutPosition;
};

/**
 * Mode in which the points are currently being deployed. If not provided during PUT/POST it will be defaulted to interpolated_palette
 */
export type SupportedGradientMode = 'interpolated_palette' | 'interpolated_palette_mirrored' | 'random_pixelated';

/**
 * Basic feature containing gradient properties.
 */
export type Gradient = {
    /**
     * Collection of gradients points. For control of the gradient points through a PUT a minimum of 2 points need to be provided.
     */
    points?: Array<Color>;
    mode?: SupportedGradientMode;
};

export type SupportedEffects = 'prism' | 'opal' | 'glisten' | 'sparkle' | 'fire' | 'candle' | 'no_effect' | 'underwater' | 'cosmos' | 'sunbeam' | 'enchant';

/**
 * Current status values the light is in regarding timed effects
 */
export type SupportedTimedEffects = 'sunrise' | 'sunset' | 'no_effect';

export type Dimming = {
    brightness: Brightness;
};

/**
 * color temperature in mirek or null when the light color is not in the ct spectrum
 */
export type Mirek = number;
// ...
export type Dynamics = {
    /**
     * Duration of a light transition or timed effects in ms.
     */
    duration?: number;
};
// ...
export type SceneMetadata = {
    /**
     * Human readable name of a resource
     */
    name?: string;
    image?: ResourceIdentifier;
    /**
     * Application specific data. Free format string.
     */
    appdata?: string;
};

export type ColorPaletteGet = {
    color?: Color;
    dimming?: Dimming;
};

export type ColorTemperaturePalettePost = {
    color_temperature?: {
        mirek?: Mirek;
    };
    dimming?: Dimming;
};

/**
 * Group of colors that describe the palette of colors to be used when playing dynamics
 */
export type ScenePalette = {
    color?: Array<ColorPaletteGet>;
    dimming?: Array<Dimming>;
    color_temperature?: Array<ColorTemperaturePalettePost>;
    effects?: Array<{
        effect?: SupportedEffects;
    }>;
};
// ...
export type ActionPost = {
    target: ResourceIdentifier;
    /**
     * The action to be executed on recall
     */
    action: {
        on?: On;
        dimming?: Dimming;
        color?: Color;
        color_temperature?: {
            mirek?: Mirek;
        };
        gradient?: Gradient;
        /**
         * Basic feature containing effect properties.
         */
        effects?: {
            effect?: SupportedEffects;
        };
        dynamics?: Dynamics;
    };
};
// ...
export type SceneRecall = {
    /**
     * When writing active, the actions in the scene are executed on the target. dynamic_palette starts dynamic scene with colors in the Palette object.
     */
    action?: 'active' | 'dynamic_palette' | 'static';
    /**
     * Transition to the scene within the timeframe given by duration
     */
    duration?: number;
    dimming?: Dimming;
};

export type ScenePut = {
    type?: 'scene';
    /**
     * List of actions to be executed synchronously on recall
     */
    actions?: Array<ActionPost>;
    recall?: SceneRecall;
    metadata?: SceneMetadata;
    palette?: ScenePalette;
    /**
     * Speed of dynamic palette for this scene
     */
    speed?: number;
    /**
     * Indicates whether to automatically start the scene dynamically on active recall
     */
    auto_dynamic?: boolean;
};